+++ etl-api/.sqlx/query-32d4ef8790cd672f7ad264113224bbf3cd8579f64afde9edec81f8580419f122.json
@@ -1,0 +1,23 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "\n        delete from app.replicators\n        where tenant_id = $1 and id = $2\n        returning id\n        ",
+  "describe": {
+    "columns": [
+      {
+        "ordinal": 0,
+        "name": "id",
+        "type_info": "Int8"
+      }
+    ],
+    "parameters": {
+      "Left": [
+        "Text",
+        "Int8"
+      ]
+    },
+    "nullable": [
+      false
+    ]
+  },
+  "hash": "32d4ef8790cd672f7ad264113224bbf3cd8579f64afde9edec81f8580419f122"
+}
