+++ etl-api/tests/integration/pipelines_test.rs
@@ -19,6 +19,7 @@ use sqlx::PgPool;
 use sqlx::postgres::types::Oid;
 use uuid::Uuid;
 
+use crate::common::database::{create_test_source_database, run_etl_migrations_on_source_database};
 use crate::{
     common::test_app::{TestApp, spawn_test_app},
     integration::destination_test::create_destination,
@@ -150,36 +151,9 @@ async fn setup_pipeline_with_source_db() -> (TestApp, String, i64, PgPool, PgCon
     .await;
 
     // We run the migrations to create all the tables used by `etl`.
-    run_etl_migrations_on_source_db(&source_db_config).await;
+    run_etl_migrations_on_source_database(&source_db_config).await;
 
     (app, tenant_id, pipeline_id, source_pool, source_db_config)
-}
-
-/// Runs etl migrations on the source database.
-async fn run_etl_migrations_on_source_db(source_db_config: &PgConnectionConfig) {
-    // We create a pool just for the migrations.
-    let source_pool = connect_to_source_database(source_db_config, 1, 1)
-        .await
-        .unwrap();
-
-    // Create the `etl` schema first.
-    sqlx::query("create schema if not exists etl")
-        .execute(&source_pool)
-        .await
-        .unwrap();
-
-    // Set the `etl` schema as search path (this is done to have the migrations metadata table created
-    // by sqlx within the `etl` schema).
-    sqlx::query("set search_path = 'etl';")
-        .execute(&source_pool)
-        .await
-        .unwrap();
-
-    // Run replicator migrations to create the state store tables.
-    sqlx::migrate!("../etl-replicator/migrations")
-        .run(&source_pool)
-        .await
-        .unwrap();
 }
 
 /// Creates a table with a chain of replication states.
@@ -272,40 +246,6 @@ async fn test_rollback(
     } else {
         None
     }
-}
-
-async fn create_test_source_database(
-    app: &TestApp,
-    tenant_id: &str,
-) -> (PgPool, i64, PgConnectionConfig) {
-    let mut source_db_config = app.database_config().clone();
-    source_db_config.name = format!("test_source_db_{}", Uuid::new_v4());
-
-    let source_pool = create_pg_database(&source_db_config).await;
-
-    let source_config = SourceConfig {
-        host: source_db_config.host.clone(),
-        port: source_db_config.port,
-        name: source_db_config.name.clone(),
-        username: source_db_config.username.clone(),
-        password: source_db_config
-            .password
-            .as_ref()
-            .map(|p| SerializableSecretString::from(p.expose_secret().to_string())),
-    };
-
-    let source = CreateSourceRequest {
-        name: "Test Source".to_string(),
-        config: source_config,
-    };
-
-    let response = app.create_source(tenant_id, &source).await;
-    let response: CreateSourceResponse = response
-        .json()
-        .await
-        .expect("failed to deserialize response");
-
-    (source_pool, response.id, source_db_config)
 }
 
 async fn create_test_table(source_pool: &PgPool, table_name: &str) -> Oid {
@@ -723,66 +663,6 @@ async fn all_pipelines_can_be_read() {
             insta::assert_debug_snapshot!(pipeline.config);
         }
     }
-}
-
-#[tokio::test(flavor = "multi_thread")]
-async fn deleting_a_source_cascade_deletes_the_pipeline() {
-    init_test_tracing();
-    // Arrange
-    let app = spawn_test_app().await;
-    create_default_image(&app).await;
-    let tenant_id = &create_tenant(&app).await;
-    let source_id = create_source(&app, tenant_id).await;
-    let destination_id = create_destination(&app, tenant_id).await;
-
-    let pipeline = CreatePipelineRequest {
-        source_id,
-        destination_id,
-        config: new_pipeline_config(),
-    };
-    let response = app.create_pipeline(tenant_id, &pipeline).await;
-    let response: CreatePipelineResponse = response
-        .json()
-        .await
-        .expect("failed to deserialize response");
-    let pipeline_id = response.id;
-
-    // Act
-    app.delete_source(tenant_id, source_id).await;
-
-    // Assert
-    let response = app.read_pipeline(tenant_id, pipeline_id).await;
-    assert_eq!(response.status(), StatusCode::NOT_FOUND);
-}
-
-#[tokio::test(flavor = "multi_thread")]
-async fn deleting_a_destination_cascade_deletes_the_pipeline() {
-    init_test_tracing();
-    // Arrange
-    let app = spawn_test_app().await;
-    create_default_image(&app).await;
-    let tenant_id = &create_tenant(&app).await;
-    let source_id = create_source(&app, tenant_id).await;
-    let destination_id = create_destination(&app, tenant_id).await;
-
-    let pipeline = CreatePipelineRequest {
-        source_id,
-        destination_id,
-        config: new_pipeline_config(),
-    };
-    let response = app.create_pipeline(tenant_id, &pipeline).await;
-    let response: CreatePipelineResponse = response
-        .json()
-        .await
-        .expect("failed to deserialize response");
-    let pipeline_id = response.id;
-
-    // Act
-    app.delete_destination(tenant_id, destination_id).await;
-
-    // Assert
-    let response = app.read_pipeline(tenant_id, pipeline_id).await;
-    assert_eq!(response.status(), StatusCode::NOT_FOUND);
 }
 
 #[tokio::test(flavor = "multi_thread")]
